Ticket delete validation has been created: To stop students from being able to delete other students' tickets buttons have been disable on the main.php file however validation has also been added. So before a ticket can be deleted first the ticket number is used to query the database to see the student number that placed the request. If the student number session variable matches that of the student number pulled from the query only then will the request be deleted.

+++ delete.php
@@ -2,17 +2,29 @@
     include('connection.php');
     session_start();
 
+    $student = $_SESSION["studentNumber"];
+
     if(isset($_POST['itemQ'])){
-        
-        $itemID = $_POST['itemQ'];
 
-        $sqlQuery = "DELETE FROM help_request WHERE TicketNo = '$itemID'";
+        $itemID = $_POST['itemQ']; //sotring ticket number in here
+        $sqlQuery = mysqli_query($connection, "SELECT StudentID FROM help_request WHERE TicketNo = '$itemID'"); //fidning student number associated with the ticket number
+        $result = mysqli_fetch_assoc($sqlQuery); //fetching the result
 
-        if (mysqli_query($connection, $sqlQuery)){
-            $response = json_encode(array('type' => 'success', 'text' => 'Item is deleted')); //message to send back to client side
-            die($response);
+        //sql query which will stop any user from deleting content - checking if students id, that is stored in the session var, matches the ticket number
+        if($student == $result["StudentID"]) {
+
+            $sqlQuery = "DELETE FROM help_request WHERE TicketNo = '$itemID'";
+    
+            if (mysqli_query($connection, $sqlQuery)){
+                $response = json_encode(array('type' => 'success', 'text' => 'Item is deleted')); //message to send back to client side
+                die($response);
+            } else {
+                $response = json_encode(array('type' => 'error', 'text' => 'Cannot delete the item')); //message to send back to client side
+                die($response);
+            }
+
         } else {
-            $response = json_encode(array('type' => 'error', 'text' => 'Cannot delete the item')); //message to send back to client side
+            $response = json_encode(array('type' => 'error', 'text' => 'Cannot do this action')); //message to send back to client side
             die($response);
         }
